pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page downloader"
authors = ["ilyas <ilyasers@gmail.com>"]
packages = [
    { include = "page_loader" },
]

[tool.poetry.scripts]
download = "page_loader.scripts.download:main"

[tool.poetry.dependencies]
python = "^3.9"
argparse = "^1.4.0"
beautifulsoup4 = "^4.10.0"
progress = "^1.6"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^7.0.0"
requests-mock = "^1.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
